+++ tests/test_ml_nn.py
@@ -14,16 +14,38 @@ from madmom.ml.nn import *
 from madmom.ml.nn.layers import *
 
 
-class TestNeuralNetworkClass(unittest.TestCase):
+class TestRNNClass(unittest.TestCase):
 
-    def test_rnn(self):
-        rnn = NeuralNetwork.load(ONSETS_RNN[0])
-        input_size = rnn.layers[0].weights.shape[0]
-        data = np.zeros((4, input_size))
-        data[1] = 1.
-        result = rnn.process(data)
-        self.assertTrue(np.allclose(result, [1.78801871e-04, 8.00144131e-01,
-                                             3.30476369e-05, 1.36037513e-04]))
+    def setUp(self):
+        # uni-directional RNN
+        self.rnn = NeuralNetwork.load(ONSETS_RNN[0])
+        self.data = np.zeros((4, self.rnn.layers[0].weights.shape[0]))
+        self.data[1] = 1.
+        self.result = [1.78801871e-04, 8.00144131e-01,
+                       3.30476369e-05, 1.36037513e-04]
+
+    def test_rnn_sequence(self):
+        # FIXME: these old models don't have the online attribute set, so we
+        #        can call either process or process_sequence
+        result = self.rnn.process(self.data)
+        self.assertTrue(np.allclose(result, self.result))
+
+    def test_rnn_step(self):
+        result_1 = np.hstack([self.rnn.process_step(d) for d in self.data])
+        self.assertTrue(np.allclose(result_1, self.result))
+        # after resetting the RNN, it must produce the same output
+        self.rnn.reset()
+        result_2 = np.hstack([self.rnn.process_step(d) for d in self.data])
+        self.assertTrue(np.allclose(result_1, result_2))
+        # without resetting it produces different results
+        result_3 = np.hstack([self.rnn.process_step(d) for d in self.data])
+        self.assertTrue(np.allclose(result_3,
+                                    [9.15636891e-04, 9.74331021e-01,
+                                     4.83996118e-05, 2.72355013e-04]))
+
+
+# class for testing all other (offline-only) networks
+class TestNeuralNetworkClass(unittest.TestCase):
 
     def test_brnn(self):
         rnn = NeuralNetwork.load(ONSETS_BRNN[0])
@@ -83,6 +105,8 @@ class TestFeedForwardLayerClass(unittest.TestCase):
         input_size = self.layer.weights.shape[0]
         self.data = np.zeros((4, input_size))
         self.data[1] = 1.
+        self.result = np.array([[0.16283005], [0.14362903],
+                                [0.16283005], [0.16283005]])
 
     def test_types(self):
         self.assertTrue(isinstance(self.layer, FeedForwardLayer))
@@ -95,11 +119,17 @@ class TestFeedForwardLayerClass(unittest.TestCase):
     def test_activate(self):
         # test result
         result_1 = self.layer(self.data)
-        self.assertTrue(np.allclose(result_1,
-                                    [[0.16283005], [0.14362903],
-                                     [0.16283005], [0.16283005]]))
+        self.assertTrue(np.allclose(result_1, self.result))
         # two runs must yield identical results
         result_2 = self.layer(self.data)
+        self.assertTrue(np.allclose(result_1, result_2))
+
+    def test_activate_step(self):
+        # test result
+        result_1 = [self.layer.activate_step(d) for d in self.data]
+        self.assertTrue(np.allclose(result_1, self.result))
+        # two runs must yield identical results
+        result_2 = [self.layer.activate_step(d) for d in self.data]
         self.assertTrue(np.allclose(result_1, result_2))
 
 
@@ -118,8 +148,8 @@ class TestRecurrentLayerClass(unittest.TestCase):
         self.assertTrue(self.layer.activation_fn == tanh)
 
     def test_init(self):
-        self.assertFalse(hasattr(self.layer, 'init'))
-        self.assertFalse(hasattr(self.layer, '_prev'))
+        self.assertTrue(hasattr(self.layer, 'init'))
+        self.assertTrue(hasattr(self.layer, '_prev'))
 
     def test_activate(self):
         # test result
@@ -131,6 +161,31 @@ class TestRecurrentLayerClass(unittest.TestCase):
         # two runs must yield identical results
         result_2 = self.layer(self.data)
         self.assertTrue(np.allclose(result_1, result_2))
+        # initialisation must not change
+        self.assertTrue(np.allclose(self.layer.init, np.zeros(25)))
+
+    def test_activate_step(self):
+        # reset layer
+        self.layer.reset()
+        result_1 = [self.layer.activate_step(d) for d in self.data]
+        result_1 = np.vstack(result_1)
+        self.assertTrue(np.allclose(result_1[0, :2],
+                                    [-0.33919713, -0.02091585]))
+        self.assertTrue(np.allclose(result_1[-1, -2:],
+                                    [0.4419672, -0.261151]))
+        self.assertTrue(np.allclose(self.layer._prev, result_1[-1]))
+        # activate without resetting the layer
+        result_2 = [self.layer.activate_step(d) for d in self.data]
+        result_2 = np.vstack(result_2)
+        self.assertTrue(np.allclose(self.layer._prev, result_2[-1]))
+        # results must be different
+        self.assertFalse(np.allclose(result_1, result_2))
+        self.assertTrue(np.allclose(result_2[0, :2],
+                                    [-3.14807342e-01, -2.22700375e-01]))
+        self.assertTrue(np.allclose(result_2[-1, -2:],
+                                    [4.40259654e-01, -2.59141315e-01]))
+        # initialisation must not change
+        self.assertTrue(np.allclose(self.layer.init, np.zeros(25)))
 
 
 class TestLSTMLayerClass(unittest.TestCase):
@@ -156,8 +211,8 @@ class TestLSTMLayerClass(unittest.TestCase):
         self.assertTrue(self.layer.output_gate.activation_fn == sigmoid)
 
     def test_init(self):
-        self.assertFalse(hasattr(self.layer, 'init'))
-        self.assertFalse(hasattr(self.layer, '_prev'))
+        self.assertTrue(hasattr(self.layer, 'init'))
+        self.assertTrue(hasattr(self.layer, '_prev'))
 
     def test_activate(self):
         # test result
@@ -169,6 +224,24 @@ class TestLSTMLayerClass(unittest.TestCase):
         # two runs must yield identical results
         result_2 = self.layer(self.data)
         self.assertTrue(np.allclose(result_1, result_2))
+        # previous step must be last result
+        # self.assertTrue(np.allclose(self.layer._prev, result_2[-1]))
+
+    def test_activate_step(self):
+        self.layer.reset()
+        result_1 = [self.layer.activate_step(d) for d in self.data]
+        result_1 = np.vstack(result_1)
+        # activate without resetting the layer
+        result_2 = [self.layer.activate_step(d) for d in self.data]
+        result_2 = np.vstack(result_2)
+        self.assertFalse(np.allclose(result_1, result_2))
+        self.assertTrue(np.allclose(result_2[0, :2],
+                                    [2.34878659e-01, -1.26488507e-03]))
+        self.assertTrue(np.allclose(result_2[-1, -2:],
+                                    [-2.45573238e-01, -2.92062219e-02]))
+        self.assertTrue(np.allclose(self.layer._prev, result_2[-1]))
+        # initialisation must not change
+        self.assertTrue(np.allclose(self.layer.init, np.zeros(25)))
 
 
 class TestGRUClass(unittest.TestCase):
@@ -197,6 +270,11 @@ class TestGRUClass(unittest.TestCase):
                    [-1.12282135, 0.3780883, 1.42017503],
                    [0.62669439, 0.89438929, -0.69354132],
                    [0.16162221, -1.00166208, 0.23579985]])
+    OUT = np.array([[0.22772433, -0.13181415],
+                    [0.49479958, 0.51224858],
+                    [0.08539771, -0.56119639],
+                    [0.1946809, -0.50421363],
+                    [0.17403202, -0.27258521]])
     H = np.array([0.02345737, 0.34454183])
 
     def setUp(self):
@@ -211,9 +289,9 @@ class TestGRUClass(unittest.TestCase):
         self.gru_1 = layers.GRULayer(self.reset_gate, self.update_gate,
                                      self.gru_cell)
         self.gru_2 = layers.GRULayer(self.reset_gate, self.update_gate,
-                                     self.gru_cell, hid_init=TestGRUClass.H)
+                                     self.gru_cell, init=TestGRUClass.H)
 
-    def test_process(self):
+    def test_activate(self):
         self.assertTrue(
             np.allclose(self.reset_gate.activate(TestGRUClass.IN[0, :],
                         TestGRUClass.H), np.array([0.20419282, 0.08861294])))
@@ -224,13 +302,13 @@ class TestGRUClass(unittest.TestCase):
             np.allclose(self.gru_cell.activate(TestGRUClass.IN[0, :],
                         TestGRUClass.H, TestGRUClass.H),
                         np.array([0.9366396, -0.67876764])))
-        self.assertTrue(
-            np.allclose(self.gru_1.activate(TestGRUClass.IN),
-                        np.array([[0.22772433, -0.13181415],
-                                  [0.49479958, 0.51224858],
-                                  [0.08539771, -0.56119639],
-                                  [0.1946809, -0.50421363],
-                                  [0.17403202, -0.27258521]])))
+        # activation of the layer
+        self.assertTrue(np.allclose(self.gru_1.activate(TestGRUClass.IN),
+                                    TestGRUClass.OUT))
+        # activating the layer a second time must give the same results
+        self.assertTrue(np.allclose(self.gru_1.activate(TestGRUClass.IN),
+                                    TestGRUClass.OUT))
+        # activating the other layer
         self.assertTrue(
             np.allclose(self.gru_2.activate(TestGRUClass.IN),
                         np.array([[0.30988133, 0.13258138],
@@ -238,6 +316,31 @@ class TestGRUClass(unittest.TestCase):
                                   [0.21366976, -0.55568963],
                                   [0.30860096, -0.43686554],
                                   [0.28866628, -0.23025239]])))
+
+    def test_activate_step(self):
+        # reset layer
+        self.gru_1.reset()
+        # first activate the layer normally
+        result_1 = [self.gru_1.activate_step(d) for d in self.IN]
+        result_1 = np.vstack(result_1)
+        self.assertTrue(np.allclose(result_1, self.OUT))
+        # the previous state must be the last output
+        self.assertTrue(np.allclose(self.gru_1._prev, self.OUT[-1]))
+        # activate with same data without resetting
+        result_2 = [self.gru_1.activate_step(d) for d in self.IN]
+        result_2 = np.vstack(result_2)
+        # results must differ
+        self.assertFalse(np.allclose(result_1, result_2))
+        self.assertTrue(np.allclose(result_2, [[0.3254016, -0.33195618],
+                                               [0.53048187, 0.51293057],
+                                               [0.12495148, -0.559039],
+                                               [0.22988503, -0.48455557],
+                                               [0.20934506, -0.25959042]]))
+        # the previous state must be the last output
+        self.assertTrue(np.allclose(self.gru_1._prev,
+                                    [0.20934506, -0.25959042]))
+        # initialisation must not change
+        self.assertTrue(np.allclose(self.gru_1.init, [0, 0]))
 
 
 class TestBatchNormLayerClass(unittest.TestCase):